TEST Test that we got the same object that we put into the database
We compare that all fields we put are already there, but not that we are not
getting more fields than we stored. A more strict approach would be testing that
we also have an _id and _rev field, but we don't think it's necessary. Any
change on id or rev reporting will break the tests anyway.